+++ sphinx/source/conf.py
@@ -27,9 +27,9 @@ author = 'Oliver Zehentleitner'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.17.0'
+release = '1.17.1'
 
-html_last_updated_fmt = "Sep 16 11:23:15 UTC 2020"
+html_last_updated_fmt = "Fr 18. Sep 21:07:02 UTC 2020"
 
 # -- General configuration ---------------------------------------------------